moved setup_table to central location
